refactor: Move all db controllers to own package
